+++ package.json
@@ -30,8 +30,8 @@
     "@react-native-firebase/remote-config": "^21.6.0",
     "@react-native-firebase/storage": "^21.6.0",
     "@react-native-picker/picker": "2.9.0",
-    "@react-navigation/native": "^5.9.8",
-    "@react-navigation/stack": "^5.14.9",
+    "@react-navigation/native": "^7.0.0",
+    "@react-navigation/stack": "^7.0.0",
     "@tanstack/react-query": "^5.55.4",
     "dayjs": "^1.10.7",
     "expo": "^52.0.0",
