+++ src/screens/RouteSearchScreen.tsx
@@ -109,9 +109,7 @@ const RouteSearchScreen = () => {
   const onPressBack = useCallback(() => {
     if (navigation.canGoBack()) {
       navigation.goBack()
-      return
     }
-    navigation.navigate('MainStack')
   }, [navigation])
 
   const handleSubmit = useCallback(() => {
@@ -189,10 +187,12 @@ const RouteSearchScreen = () => {
           ...prev,
           stations: route?.stops ?? [],
           selectedDirection: direction,
-          selectedBound: route?.stops[route?.stops.length - 1] ?? null,
+          selectedBound:
+            direction === 'INBOUND'
+              ? route?.stops[route.stops.length - 1] ?? null
+              : route?.stops[0] ?? null,
         }))
-        setNavigationState((prev) => ({ ...prev, trainType: null }))
-        navigation.navigate('Main')
+        navigation.goBack()
         return
       }
 
@@ -200,9 +200,8 @@ const RouteSearchScreen = () => {
         lineGroupId: trainType.groupId,
       })
 
-      const station = data.stations.find(
-        (s) => s.groupId === currentStation?.groupId
-      )
+      const station =
+        data.stations.find((s) => s.groupId === currentStation?.groupId) ?? null
 
       const direction: LineDirection =
         data.stations.findIndex((s) => s.groupId === currentStation?.groupId) <
@@ -213,9 +212,11 @@ const RouteSearchScreen = () => {
       setNavigationState((prev) => ({
         ...prev,
         trainType: station?.trainType ?? null,
+        stationForHeader: station,
       }))
       setStationState((prev) => ({
         ...prev,
+        station,
         stations: data.stations,
         selectedDirection: direction,
         selectedBound:
@@ -223,10 +224,10 @@ const RouteSearchScreen = () => {
             ? data.stations[data.stations.length - 1]
             : data.stations[0],
       }))
-      navigation.navigate('Main')
+      navigation.goBack()
     },
     [
-      currentStation?.groupId,
+      currentStation,
       fetchTrainTypeFromTrainTypeId,
       navigation,
       selectedStation?.groupId,
